requirements.txt
openai>=1.40.0
chromadb>=0.5.5
python-dotenv>=1.0.1
tiktoken>=0.7.0
rich>=13.7.0
streamlit>=1.36.0
